backend/requirements.txt
aniso8601==10.0.1
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
exceptiongroup==1.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-flask==1.3.0
python-dotenv==1.2.1
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.4
